+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.8.1]() - 2024-02-23
+## [3.9.0]() - 2025-03-03
 ### Fixed
 - Fixed paging when filtering Get-OMEDevice by Group
 - Fixed paging in Get-OMEJob
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Get-OMEDiscovery
+- Invoke-OMEBlinkLED
 
 ## [3.8.0]() - 2024-02-23
 ### Fixed